+++ examples/arenera.jl
@@ -41,12 +41,14 @@ function testRun(seed::Int64)
 
     # @profview , @time
     model, output = jOpqua.simulate!(
-        model, t_vec, host_samples_population=Dict("pop"=>100)
+        model, t_vec, host_samples_population=Dict("pop" => 100)
     )
-    println(output.compartment_vars["pop"][:,end])
-    println(output.host_samples["pop"][:,end][1:3])
+    println(output.compartment_vars["pop"][:, end])
+    println(output.host_samples["pop"][:, end][1:3])
 
     jOpqua.plotCompartments(output, ["pop"], "examples/compartment_test.png")
+    jOpqua.saveCompartments(output, "examples/compartment_test.csv")
+    jOpqua.saveHistory(output, "examples/history_test.csv")
 end
 
 @time testRun(1)